feat/7/fix: ERD와 코드 통일 (#8)
ERD와 코드 통일

+++ back/src/main/java/csh/back/domain/trip/timeline/dto/request/TimelineCreateRequest.java
@@ -13,7 +13,7 @@ public record TimelineCreateRequest(
         @Schema(description = "여행 일차 번호", example = "1")
         @NotNull
         @Min(1)
-        Integer dayNumber,
+        Long dayNumber,
 
         @Schema(description = "타임라인 시작 시간", example = "2026-07-01T09:00:00")
         @NotNull
